Improve liveness probes
The current liveness probes for Cinder Backup, Scheduler, and Volume
services use a bash command (templates/cinder/bin/healthcheck.sh) to
check if the service is up.

This has the following downsides:

- The cinder-api service needs to be up and running

- Need connectivity to the cinder-api service

- They are slow, since it does 7 crudini calls, one to the cinder client
  (which in turn makes 1 call to keystone and 2 to cinder), one to grep,
  and one to awk.

- It's not ideal for containers to be constantly running commands.

This patch replaces the liveness probes with HTTP probes that are served
by a very simple HTTP server on a sidecar container.

This HTTP Server is `templates/cinder/bin/healthcheck.py` that takes the
name of the service (volume, backup, or scheduler) and optionally the
location of the configuration file (defaults to
/etc/cinder/cinder.conf), reads the service configuration (including the
DB connection bits) and then, when it receives a GET request, queries
the services table in the cinder database and checks that the service is
up.

The Server is currently able to differentiate between 3 different kind
of errors:

- Container not being able to reach the database
- Service not present in the database table
- Service present but appears as down

For the volume service it will only consider that the service is OK if
ALL the `enabled_backends` are up.

In the future it could be extended to do a RabbitMQ query to the service
to confirm RabbitMQ connectivity.